configurator: libvirt: do not jump on virtlogd
Libvirt 1.3.0 adds and uses by default a new service called virtlogd
to prevent malicious QEMU guests to DoS the virtualization host.
For more details, the initial announcement is here:
https://www.redhat.com/archives/libvir-list/2015-November/msg00085.html

The feature seems promising and helpful, but we are
not eager to jump early on this bandwagon, do we disable
this for the near future.

Once the feature is stable and we run our integration test,
we will re-enable it.

Change-Id: I548560af8e78a69210917c36d229206663beaaf7
Backport-To: 3.6
Temporary-Patch: yes
Bug-Url: https://bugzilla.redhat.com/1318902
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/55189
Continuous-Integration: Jenkins CI
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/55225