blockVolume: Remove stale links to image directory
Since we started to create the image directory under "/run/vdsm", the
image directory under "/rhev/data-center/pool/sd/images/" may be a link
to "/run/vdsm/storage/sd/image".

The old code (from rhev-3.0) creating the image directory lazily was not
designed to handle links and was happily using a stale link instead of
creating the image directory.

Now we detect and remove stale links, so proper image directory is
created. Since stale links are not expected in normal usage, we log a
warning in this case.

Change-Id: Ic694caab0e030c675e4a3a0c0638298abfea1a8c
Bug-Url: https://bugzilla.redhat.com/1319987
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/55175
Continuous-Integration: Jenkins CI