net: Libvirt hook that enables ovs-legacy migration
This hook allows VM/s to be migrated between OVS based hosts and legacy
based hosts, in both directions.

The hook is added to the OVS hook package.

Note: The OVS hook and this migration hook is limited to a single
non-vlan network. An attempt to migrate a VM to a host with multi
non-vlan networks will fail.

Change-Id: I29cf441cc365d3679382e44410dad0906d9be3ec
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/55497
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>