fix 186c8574: When using SSL, use the original host
When a connection to a remote is perform and encryption (SSL) is
enabled, the original target host should be used for the connection and
not a resolved one.

This bug has been detected when trying to connect to a remote IPv4 host,
with a hostname as the target host and SSL enabled.
The host got resolved before calling connect, and therefore the
certificate has not been matched correctly (expected a hostname but used
a resolved IPv4 address).

When M2C is dropped, we can easily use the server_hostname arg of
https://docs.python.org/2/library/ssl.html#ssl.SSLContext.wrap_socket
and avoid the duplicate name resolution.

Change-Id: Ibd54bcddd5ab676d94f7a5965a061d1a3cb1f40a
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/55620
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>