migrations: change migration parameters
Changed the migration params to:
max_outgoing/incoming_migrations=2
migration_max_bandwidth=52
migration_progress_timeout=240
migration_downtime_delay=20
migration_downtime_steps=5

The reason to change migration_max_bandwidth and
max_outgoing/incoming_migrations is to have less migrations running in parallel
faster which helps convergence.

The reason to change the migration_progress_timeout is to leave more time for
big VMs to converge.

The reason to change migration_downtime_delay is to make the downtime thread
start applying the changes faster which is important for VMs which would start
stalling.

The reason to change the migration_downtime_steps is to make the growing of the
downtime more fast.

Change-Id: I1ed1b04ee77a7144ae0206e5c5ae1426ede0c75f
Bug-Url: https://bugzilla.redhat.com/1323952
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/55676
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>