hsm: add untrusted image verification.
In some flows, where an image can be imported into oVirt's storage from
external sources, e.g in the image upload flow- a VDSM's image is being
generated and data is being sent into the created volume inside.

As this flow is risky in cases where the data is a QCOW file with an
invalid backing file, or wrong disk format is sent, VDSM should verify
the image in the upload process.

This patch adds a verb for untrusted image's volume verification, which
verifies that there is no backing file for a QCOW image, and that the
format of the uploaded image fits to what the user has specified.

Change-Id: Ibf85061536eb4ddff021539c742a674f183a8984
Signed-off-by: Amit Aviram <aaviram@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/55746
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI