virt: Make DomainDescriptor use XML helpers
We are going to stop using xml.dom.minidom and to use xml.etree instead.
XML processing is currently spread across several places and we use both
minidom and etree, so we must be careful with the conversion.  The
overall strategy is as follows:

1. Create XML helpers abstracting common XML operations.
2. Use the helpers everywhere where minidom is currently used.  Avoid
   using minidom anywhere outside the XML helpers.
3. Switch from minidom to etree in the XML helpers.

We avoid using XML directly in many places as a side effect.  It is
debatable whether it is a good or bad thing.  If we decide it is a good
thing, we may want to replace direct etree base XML processing with the
helpers as well in a followup patch series.  If we decide it is a bad
thing, we can unfold the helpers in a followup patch series (note we
need the helpers as an intermediate step and can't avoid them from the
beginning).  Keeping status quo after this series is also an option.

There's also duplicate XML code and processing present in some places,
especially in core devices.  We keep it duplicated as it is in this
series, in order not to mix etree conversion with refactoring.  Device
XML processing refactoring is planned in a followup patch series.

This patch implements the first step of the whole process.  It converts
DomainDescriptor class from minidom to XML helpers.  The reason for this
start is that DomainDescriptor class is simple while it requires most of
the needed XML helpers.  So we can start with a small change of the
existent code while becoming equipped with a nice set XML helpers for
the following, perhaps more complicated, patches.

Change-Id: Ib169735936d19171ff8b8d127666d7258c308f34
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/55769
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI