schema: SDM.create_volume split parent params
When creating a new volume, the parent_vol_id and parent_img_id are
optional and should be used only if the new volume is derived from a
parent (ie. snapshot volume or volume from template).  However, if one
of the parent_*_id parameters is specified the other one must also be
provided.  The best way to codify this and organize the create
parameters is to place the two parent ids together in their own data
structure.

Change-Id: I7ccaf88a122dec24fa8d018c0cd94b4a08ecadce
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/55829
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Freddy Rolland <frolland@redhat.com>