schedule: Port schedule module to Python 3
In Python 3, you cannot use push a tuple into a heap if one of the
objects in the tuple does not support ordering. Implement the minimal
rich comparison to make it work on Python 3.

Since ScheduledCall supports ordering now, we don't need to wrap it in a
tuple, and we can push the object itself into the heap. We used a tuple
before because in Python 2 it was faster than implementing a __cmp__
method. Unfortunately, we don't have this option with Python 3.

Change-Id: Ie6d20fe1a1d262d5489be5bedba929709fa59e35
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/55880
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>