build: Make sure run_tests*.sh scripts are executable
Generated script tests/run_tests_local.sh is not executable after
`make clean; make' is run.  This bug was introduced by commit 6e6f89b7,
but there is probably nothing wrong with that commit, it just introduces
removal of some files including run_tests_local.sh.  The problem is
tests/run_tests*.sh files are not set as executable if they are
recreated any time later.

This patch tries to solve the problem by making the following changes:

- tests/run_tests*.sh are no longer handled in configure.ac.  They are
  not config files and such a use of AC_OUTPUT is obsolete anyway.
- New tests/Makefile.am rule is introduced ensuring generated
  run_tests*.sh files are set as executable.
- top_srcdir substitution is added to Makefile.subs to make the inserted
  relative paths in the generated files always correct.
- run_tests_local.sh is moved to *_SCRIPTS target in tests/Makefile.am.

There may be other solutions of the problem but whatever solution is
chosen, it should be checked that tests/run_tests*.sh files are
executable and the paths inside the generated files are right in all the
following situations:

- After running ./autogen.sh + make in a fresh git directory.
- After `rm tests/run_tests*.sh; make'.
- After `(cd tests; rm run_tests*.sh; make)'.

Change-Id: Ibc1e3dc8ace7f69801b765262352903020cc8aef
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/55949
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Tomas Golembiovsky <tgolembi@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>