hostdev: report additional information in 'scsi' device
SCSI capability in sysfs does not include enough information to
identify the device in human readable way. We therefore fetch vendor
and product from 'storage' capability, where these correspond to a
physical device (if there is no physical device upstream, we ignore it).

Another addition is udev path, that we require to fully construct
the device in future.

Change-Id: I789f49c9abca2dff1487acf3e8a04ae1407956f4
Bug-Url: https://bugzilla.redhat.com/1270581
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56038
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>