virt: storage: avoid empty serials to domain XML
The VM drives in the libvirt domain XML may have
one 'serial' attribute.
Vdsm used to always sends this attribute.
The value is either received as drive configuration
parameter, or it is filled from the drive backing imageID.

Should Vdsm not receive the serial attribute, or should
the drive not have a backing image, the serial will be
empty, and will be sent anyway to libvirt.

Starting from libvirt 1.3.0, libvirt enforces the fact
that the serial must not be empty (see
https://bugzilla.redhat.com/1322842#c4).
This exposed an old bug of Vdsm, and the effect is that
libvirt rejects the domain XML and the VM could not start.

To trigger this bug, just configure one VM with payload
device (e.g. cloud init) this will send a new cdrom device
alongside the default one, and both will have empty serial.

Looking again at the libvirt domain XML specifications
(http://libvirt.org/formatdomain.html#elementsDisks)
we see that the 'serial' attribute is not mandatory.

Thus, the simplest fix is just not to add the 'serial'
attribute in the domain XML if it is empty.

Change-Id: I175954321faef8fb33ee31b6b367bfad9ec87137
Bug-Url: https://bugzilla.redhat.com/1322842
Backport-To: 3.6
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56014
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-on: https://gerrit.ovirt.org/56274