qemuimg: Validate qcow2_compat configuration
Validate new qcow2_compat option against a list of approved values, and
raise (new) InvalidConfiguration error if the value is not supported:

   InvalidConfiguration: Invalid configuration value: ("Unsupported value
   for irs:qcow2_compat: '1.2'",)

This document the ones we accept and help to diagnose configuration
mistakes. We need to maintain this list over time, but it's good for us
to check new compat versions to see if we're ready to allow them to be
used in vdsm storage.

Change-Id: I8e217ec42a803d53403947a5fb140cb51202fbfa
Bug-Url: https://bugzilla.redhat.com/1317850
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56110
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-on: https://gerrit.ovirt.org/56523
Reviewed-by: Francesco Romani <fromani@redhat.com>