spec: qemu-kvm-img to handle unaligned images
Bug #1282239 describes a bug in qemu-img convert's ability to handle
exporting/importing of memory volumes, that was worked around by using
dd instead.

This patch requires a qemu-img version that fixes the underlying issue
in order to handle importing previously exported volumes and possibly
remove this workaround in the future.

Change-Id: I411b14cfc9eb72ceebb005284a95d316bba783c6
Bug-Url: https://bugzilla.redhat.com/1283278
Related-To: https://bugzilla.redhat.com/1282239
Signed-off-by: Allon Mureinik <amureini@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56142
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>