build: Remove pep8 and pyflakes requirement
We use to build-require both pep8 and pyflakes, which cause trouble when
using older version. This leads to build failures, or to horrible code
tyring to make old buggy pyflakes happy.

Vdsm does not require pep8 or pyflakes for building packages; these are
development tools that should be installed on a development machine or
on CI environment, using the versions supported by current vdsm code.

In the makefile, "make check" during make rpm was disabled on EL,
because incompatible or missing pep8 and pyflakes versions; This hack
was removed, so make rpm will do the same thing all platforms.

In the makfile, we run pep8 or pyflakes checks only if these tools are
installed.

4 copies of incompatible pep8 and pyflakes version were removed from the
packages list.

Finally, we install the correct version of these tools in the CI
environment using pip.

Change-Id: I6f0cdf21f18604c3fe81fa724aab0e613fb04289
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56152
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Tested-by: Yaniv Bronhaim <ybronhei@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>