tests: storagetestlib: normalize make_file_volume
The test utility make_block_volume was recently improved so that the
created volume had correct metadata in the fake storage domain.  Do the
same for make_file_volume and normalize the interface so that it works
like make_block_volume (ie. requires image and volume ids).

Change-Id: I79a94da73e370541277ad6fa55fcf4655681e1c2
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56170
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI