asyncore: make sure to handle sslerror on send
When we receive data we handle ssl errors in cases like a client
disconections. We want to make sure that we close connection and return
to message processing code to handle the issue.

In send we have similar behavior but the error propagates to asyncore
loop where it is handled. For send we need to have similar behavior:
- close channel
- let the sender code handle case where no data was send


Change-Id: I28fd1accb04c74c8ed2f11ba47fa5ef7fee67883
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: https://gerrit.ovirt.org/56206
Continuous-Integration: Jenkins CI
Tested-by: Simone Tiraboschi <stirabos@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>