net: Clean running config on unified restoration failure
When the unified restoration process fails, running config (file) may
contain networks/bonds that are not in kernel.
When an attempt is made to re-run the unified restoration (without a
boot), the process will fail to remove the mentioned networks/bonds.

In this patch, the running config will be cleared if the restoration
fails, allowing the following attempts to try and setup the network
again.

Change-Id: Ie43da8ca26ef0b0e92ec66e1a481a843aea0487f
Bug-Url: https://bugzilla.redhat.com/1323782
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56257
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>