NetworkManager: configure to monitor ifcfg/connection files
The new configuration file instructs NetworkManager to automatically
reload network configuration from the directory it uses to persist
its "connections" (one could say "profiles"):

- /etc/sysconfig/network-scripts/ in the case of ifcfg systems
  (Fedora, RHEL, CentOS and derivatives; "ifcfg-rh" plugin)
- /etc/NetworkManager/system-connections/ ("keyfile" plugin)
- or another, specific to e.g. Debian or Suse

This functionality is required to properly let NetworkManager
know when VDSM consumes any of its devices (on ifcfg systems,
by writing NM_CONTROLLED=no to an ifcfg file).

Please note that this approach is potentially raceful. Under
high loads, it is theoretically possible that NM learns of
a changed file too late. To be sure, we should probably call
to 'nmcli connection load' synchronously, just before running
'ifup' on a given device.

NetworkManager is asked to reload its configuration during
the %post stage of installation. If it is not running,
systemctl simply exits with exit code 130 and doesn't try
to start it.

Change-Id: I3eb80b3a280abd72c7084860467aca12907d0477
Bug-Url: https://bugzilla.redhat.com/1304509
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56365
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>