virt: Set timeout on boot menu
The default timeout for boot menu (3 s) is too short for some scenarios,
e.g. with network delays.  Although the best way to handle boot menu is
to run a VM in paused mode, we can change the timeout to make boot menu
more available even on regular runs.  So this patch sets the boot menu
timeout to 10 s, which could be a reasonable value, not too low and not
too high.

Change-Id: I0f3501e8500e366e785f5a8ddfdf78fd34c997a2
Bug-Url: https://bugzilla.redhat.com/1324375
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56393
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI