sdm: VolumeArtifacts COW base volume support
Add support for qcow2 base volumes.  This is done by using the qemu-img
program to initialize the prepared volume.  The rest of the patch is
devoted to testing corner cases and invalid parameters.

This adds support for base volumes only.  Snapshots and Templates will
be supported in future patches.

Change-Id: If31e013187830322d4c9ebf0d76b4d6d4d81445d
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56629
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI