Support initial_size parameter
The non-SDM volume creation API supports an initial_size parameter which
allows you to indicate the initial size of a thinly-provisioned block
volume.  We need to support this functionality in the sdm create volume
flow as well.

Note that initial_size is only legal for COW formatted block volumes.
Since we do not support those yet this patch only supports passing the
parameter and throwing errors when appropriate.  The implementation will
come along with COW volume support.

Change-Id: I795d5c8e187252c95764ede998447c6524b37f91
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56628
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>