mount: Wait for systemd temporary mount unit
Systemd is monitoring mountpoints created during runtime by creating a
temporary mount units. For example:

    # systemctl status /tmp/mount-1.DUNxySxI2R/wrk
    ● tmp-mount\x2d1.DUNxySxI2R-wrk.mount - /tmp/mount-1.DUNxySxI2R/wrk
       Loaded: loaded (/proc/self/mountinfo)
       Active: active (mounted) since Sun 2016-04-24 22:37:55 IDT; 14ms ago
        Where: /tmp/mount-1.DUNxySxI2R/wrk
         What: /dev/loop2

Mounting and unmounting quickly as we do in the tests, or when creating
a floppy image may confuse systemd, and lead to various errors:

- mountpoint is not mounted when trying to access files or unmount
- mountpoint is busy when trying to unmount

These are probably systemd or mount bugs. Until these issues are
resolve, we need to find a workaround to avoid failures in the tests or
when creating a floppy images.

This patch add udevadm.settle() call after successful mount and umount
operations.

Change-Id: I9b0318eb534f781f127882983527e1a221820478
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56544
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>