tests: Add mount.isMounted() timing tests
Add timing tests to understand how much time it takes to check if a
mountpoint is mounted when having lot of mounts. This used to be a
bottleneck on EL6 when having lot of file storage domains.

Here are the results on i7-4770 CPU @ 3.40GHz:

mountTests.TestIsMountedTiming
    test_is_mounted(1)             1 mounts: 0.000068 seconds
    test_is_mounted(50)           50 mounts: 0.000480 seconds
    test_is_mounted(100)         100 mounts: 0.000927 seconds
    test_is_mounted(1000)       1000 mounts: 0.008862 seconds

Running the tests:

    ./run_tests_local.sh mountTests.py --enable-stress-tests -s

Change-Id: Ib84396450abaa17fee402d024e0ae99c36b68ce4
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56565
Reviewed-by: Idan Shaby <ishaby@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>