v2v: Detect VM with snapshots
virt-v2v cannot properly handle conversion of VMware machine with
snapshots. Engine needs a way to know which machines have snapshots and
which do not to filter them out and/or notify the user.

Note: The snapshot related API is not yet implemented in libvirt Xen
driver. It is missing in all libvirt version to date (v1.3.4).

Change-Id: I9aa4de2faff92625cd0de8e3ae2a10a2d58aa823
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56574
Continuous-Integration: Jenkins CI
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>