require NetworkManager-config-server
Request NetworkManager not to approriate newly-created network devices.
With this configuration on host, vdsm can create bonding and vlan
devices without them taken down by NM.

This patch implictly add NetworkManager as a vdsm requirement, but this
is not a very big deal as NetworkManager is ubiquitous.

Bug-Url: https://bugzilla.redhat.com/1326798
Change-Id: I22d5a05b3d3d9273adc8301696bac3abed0d046f
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56622
Continuous-Integration: Jenkins CI
Reviewed-by: Fabian Deutsch <fabiand@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Tested-by: Fabian Deutsch <fabiand@redhat.com>