lvm: Use a constant to define default vg extent size
The default vg extent size has so far been defined as "128m" by means of
a default argument value in lvm.createVG.  In order to test LVM we need
to know this value progmatically.  Define VG_EXTENT_SIZE_MB in
storage.constants and eliminate the parameter in lvm.createVG which is
not being used.

Change-Id: I77d76e9c1011cb3bba41decbcb06830eaed66a6a
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56715
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI