tests: Fix volume size handling in domain_manifest_test
This test module is not testing the rounding of volume sizes up to the
VG extent size (128M).  That is already tested in storagefakelibTests
and storage_volume_artifacts_test.  For this test, just use a large
enough size that will not need to be rounded.

Also, fix test_metaslot_selection which was not converting the volume
size to MB when calling the lvm createLV API.

Change-Id: I30a5d331585e916f3cfe0e698e1e30fb25539863
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56820
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI