virt: Don't fail when existingConnAction is unset for a SPICE device
When a ticket for a SPICE device is set, existingConnAction parameter is
sent by Engine as well.  This parameter is used to set `connected'
attribute of <graphics> element.  But setting this parameter when
updating the ticket is basically useless.  It may be actually even
harmful, e.g. when a user sets `connected' attribute value to `keep' in
a Vdsm hook and Engine overrides it to `disconnect' without any good
reason.  See the referenced bug.

This should be fixed on the Engine side.  One approach may be not to
send the parameter at all when updating the SPICE ticket.  However, Vdsm
crashes in such a case.  This patch fixes Vdsm problems when the
parameter is not present, making future changes in Engine possible.

Change-Id: I2fde83ceb994eaafd2aa956d1fa82ce6cb16094c
Bug-Url: https://bugzilla.redhat.com/1060573
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56836
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>