monitor: Detach read delay from storage domain
We used to call StorageDomain.getReadDelay() to perform a read delay
check every 10 seconds. Because the domain object is a proxy, accessing
readDealy can block for unlimited time if the real domain object is not
int the sdCache cache, blocking on the refreshStorage sampling method.

This can lead to blocking of file based storage domain monitoring
because the thread holding the refreshStorage lock is performing a
scsi/fc scan.

This patch moves getting read delay into the monitor thread. Once we
produce a domain object, we get and store the domain monitoring path.
From this point, performing a read delay does not access the domain
object and can block only if this particular storage domain is not
accessible.

Change-Id: I1a687f70c220a86e538bf1cee500aa8b6028eff1
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56826
Reviewed-by: Adam Litke <alitke@redhat.com>
Continuous-Integration: Jenkins CI