netinfo: improve the detection if device has VLAN devices on top.
The current code assumed that the VLAN device respects the naming
convention of nameOfUnderlyingDevice.VlanId. Since this convention is
not enforced, the detection fails if the device has VLAN devices on top,
which do not respect this naming convention.

The improvement is that now we will check if the device is the used
device by any link of the netlink library.

Change-Id: If55606d8313d73bfe9501e976c4335080e59f994
Signed-off-by: Dominik Holler <dominik.holler@gmx.net>
Reviewed-on: https://gerrit.ovirt.org/56858
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI