virt: add kernel cmdline reporting
As we want to control kernel command line from ovirt, we need a way to
let users (and us) see if the changes are successfully applied. We
approach this by reporting kernel command line from /proc/cmdline
without any modifications, allowing for relatively easy debugging.

Name kernelArgs was chosen as we already refer in the same way to guest
kernel arguments.

Also, using this patch as an opportunity to boostrap osinfo tests.

Change-Id: I9cfb9bef65391b9ae6fff7abeb9da64408f90ad5
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1301104
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/57084
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>