net: report OVS netinfo in caps
Even non-root caps (legacy networks+nics) are now obtained
from supervdsm. The reason is, that we do not want to keep logic
of legacy-OVS netinfo merging in two places (caps and netswitch).

When we manage to split legacy netinfo to legacy networks and NICs
parts, we will be able to simply divide netinfo between root and
non-root.

Change-Id: Ib53200891e292a0d9c8a347cc14344281d3ace65
Bug-Url: https://bugzilla.redhat.com/1195208
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/57102
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI