v2v: get_external_vms return format for KVM images
the format is needed for importing from kvm.
when importing xen/vmware we are using virt-v2v which use qemu-img to
convert the image to specific format that the user chosen.
in kvm mode we don't need the conversion we are just copying the image
with its original format.

Change-Id: I4d4985aafc21e5ff5c11ec7bca71ab9aa5b2cd9d
Signed-off-by: Shahar Havivi <shaharh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56912
Tested-by: Shahar Havivi <shavivi@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>