storage: Do not reference image dir in backing volume path
Until now, the backing volume path (to be written into qcow2 metadata)
was computed as ../<imgUUID>/<volUUID>.  The ../<imgUUID> component
causes the backing_chain field of volumes to grow every time a live
merge is performed until it eventually overflows.  The blockSD function
createImageLinks is now ensuring that all volumes of an image (including
templates) are in a directory together so we can use just the volume ID
to reference the backing volume.  File domains hard link templates into
the directories of images that are dependent so this works for file
domains as well.

Change-Id: I62afa6db6f458d5c2c64ccc3427831cb76af1dcc
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1333627
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/57284
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI