virt: Added 'prepared' context manager
Added context manager to prepare a group of objects (for example, disk
images) for an operation. Each images must implement prepare() and
teardown() methods.

When entering the context, all images are prepared in the order they are
passed. When exiting the context (or upon error) the images are torn
down in reverse order. If an exception was raised inside the context,
that exception will be propagated regardless of any exceptions that may
be thrown by teardown() methods.

Usage example:

with prepared(images):
    do_some_operation(images)

Change-Id: Ibe9d04599dfc1700f215a7765775d60cb314ff0a
Bug-Url: https://bugzilla.redhat.com/1335642
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>