cdrom: API change: require interface & index
Previously, changeCD call did not require anything but path to an
image. This works fine as long as there is only single type of
cdrom for one architecture (and set number of cdroms & architectures).

As we slowly move towards using q35 machine type, we have to plan how
to fix the cdrom since q35 does not support IDE bus. Solution that
allows different cdrom buses is requiring interface and index, as these
can uniquely identify cdrom in a system.

Change-Id: I178c1a02bbad962f9dc9b67bed7691cf170ee896
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/56805
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Arik Hadas <ahadas@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>