tests: Prevent multiple invocations of makecerts.sh
Wrapped the invocation of `makecerts.sh` in special rule to prevent
multiple concurrent invocations of the script when make is running with
multiple jobs (-j).

Also there was no rule to build `server.p12` because it was missing from
`makecerts.sh` targets in the build rule. To fix this and improve the
readability all files are specified only on one place in a variable.

Change-Id: I5553057e0c0a66ccb96aa099a6608002f42d513f
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/57344
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Martin Polednik <mpolednik@redhat.com>